+++ tests/cli/config.rs
@@ -1,4 +1,4 @@
-// Copyright 2022 Martin Pool.
+// Copyright 2022, 2023 Martin Pool.
 
 //! Test handling of `mutants.toml` configuration.
 
@@ -163,13 +163,14 @@ fn list_with_config_file_regexps() {
         "#,
     );
     run()
-        .args(["mutants", "--list", "-d"])
+        .args(["mutants", "--list", "--line-col=false", "-d"])
         .arg(testdata.path())
         .assert()
         .success()
-        .stdout(predicates::str::diff(
-            "src/simple_fns.rs:17: replace divisible_by_three -> bool with false\n",
-        ));
+        .stdout(predicates::str::diff(indoc! {"\
+                src/simple_fns.rs: replace divisible_by_three -> bool with false
+                src/simple_fns.rs: replace == with != in divisible_by_three
+        "}));
 }
 
 #[test]
@@ -178,8 +179,8 @@ fn exclude_re_overrides_config() {
     write_config_file(
         &testdata,
         r#"
-        exclude_re = [".*"]     # would exclude everything
-        "#,
+exclude_re = [".*"]     # would exclude everything
+"#,
     );
     run()
         .args(["mutants", "--list", "-d"])
@@ -189,14 +190,15 @@ fn exclude_re_overrides_config() {
         .stdout(predicates::str::is_empty());
     // Also tests that the alias --exclude-regex is accepted
     run()
-        .args(["mutants", "--list", "-d"])
+        .args(["mutants", "--list", "--line-col=false", "-d"])
         .arg(testdata.path())
         .args(["--exclude-regex", " -> "])
         .args(["-f", "src/simple_fns.rs"])
         .assert()
         .success()
         .stdout(indoc! {"
-            src/simple_fns.rs:7: replace returns_unit with ()
+            src/simple_fns.rs: replace returns_unit with ()
+            src/simple_fns.rs: replace == with != in divisible_by_three
         "});
 }